requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asyncio==3.4.3
attrs==24.3.0
boto3==1.35.90
botocore==1.35.90
certifi==2024.12.14
charset-normalizer==3.4.1
dnspython==2.7.0
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
nanoid==2.0.0
openpyxl==3.1.5
pillow==11.1.0
propcache==0.2.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
resend==2.5.1
s3transfer==0.10.4
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
